Move LiveDataUtil to the only place it's used
