+++ feature-proxy-manager/src/test/java/com/kinandcarta/create/proxytoggle/manager/LiveDataTestUtil.kt
@@ -1,4 +1,4 @@
-package com.kinandcarta.create.proxytoggle.lib.core
+package com.kinandcarta.create.proxytoggle.manager
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.Observer
